+++ src/main/java/louie/hanse/shareplate/core/entry/service/EntryService.java
@@ -4,18 +4,18 @@ import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import louie.hanse.shareplate.core.chatroom.domain.ChatRoom;
-import louie.hanse.shareplate.core.chatroom.domain.ChatRoomMember;
-import louie.hanse.shareplate.core.entry.domain.Entry;
-import louie.hanse.shareplate.core.member.domain.Member;
-import louie.hanse.shareplate.core.member.service.MemberService;
-import louie.hanse.shareplate.core.notification.domain.ActivityType;
-import louie.hanse.shareplate.core.notification.event.ActivityNotificationRegisterEvent;
-import louie.hanse.shareplate.core.share.domain.Share;
 import louie.hanse.shareplate.common.exception.GlobalException;
 import louie.hanse.shareplate.common.exception.type.EntryExceptionType;
+import louie.hanse.shareplate.core.chatroom.domain.ChatRoom;
+import louie.hanse.shareplate.core.chatroom.domain.ChatRoomMember;
 import louie.hanse.shareplate.core.chatroom.repository.ChatRoomMemberRepository;
+import louie.hanse.shareplate.core.entry.domain.Entry;
+import louie.hanse.shareplate.core.entry.event.EntryCancelEvent;
+import louie.hanse.shareplate.core.entry.event.EntryEvent;
 import louie.hanse.shareplate.core.entry.repository.EntryRepository;
+import louie.hanse.shareplate.core.member.domain.Member;
+import louie.hanse.shareplate.core.member.service.MemberService;
+import louie.hanse.shareplate.core.share.domain.Share;
 import louie.hanse.shareplate.core.share.service.ShareService;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
@@ -53,8 +53,7 @@ public class EntryService {
         ChatRoomMember chatRoomMember = new ChatRoomMember(member, chatRoom);
         chatRoomMemberRepository.save(chatRoomMember);
 
-        eventPublisher.publishEvent(
-            new ActivityNotificationRegisterEvent(shareId, memberId, ActivityType.ENTRY));
+        eventPublisher.publishEvent(new EntryEvent(shareId, memberId));
 
         return entry.getId();
     }
@@ -81,8 +80,7 @@ public class EntryService {
         chatRoomMemberRepository.deleteByMemberIdAndChatRoomId(
             memberId, share.getEntryChatRoom().getId());
 
-        eventPublisher.publishEvent(
-            new ActivityNotificationRegisterEvent(shareId, memberId, ActivityType.ENTRY_CANCEL));
+        eventPublisher.publishEvent(new EntryCancelEvent(shareId, memberId));
     }
 
     public List<Long> getIdList(Long memberId) {
