+++ src/main/java/louie/hanse/shareplate/core/notification/event/activity/ActivityNotificationRegisterEvent.java
@@ -1,4 +1,4 @@
-package louie.hanse.shareplate.core.notification.event;
+package louie.hanse.shareplate.core.notification.event.activity;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
