+++ src/main/java/louie/hanse/shareplate/core/notification/event/keyword/listener/KeywordNotificationEventListener.java
@@ -1,15 +1,14 @@
-package louie.hanse.shareplate.core.share.event.handler;
+package louie.hanse.shareplate.core.notification.event.keyword.listener;
 
 import lombok.RequiredArgsConstructor;
-import louie.hanse.shareplate.core.notification.event.ActivityNotificationRegisterEvent;
-import louie.hanse.shareplate.core.notification.event.ShareRegisterEvent;
 import louie.hanse.shareplate.core.notification.service.NotificationService;
+import louie.hanse.shareplate.core.share.event.ShareRegisterEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class ShareRegisterEventHandler {
+public class KeywordNotificationEventListener {
 
     private final NotificationService notificationService;
 
@@ -18,9 +17,5 @@ public class ShareRegisterEventHandler {
         notificationService.saveKeywordNotificationAndSend(event.getShareId(), event.getMemberId());
     }
 
-    @EventListener
-    public void registerActivityNotification(ActivityNotificationRegisterEvent event) {
-        notificationService.saveActivityNotificationAndSend(event.getShareId(), event.getMemberId(),
-            event.getActivityType());
-    }
 }
+
